+++ src/sentry/api/endpoints/integrations/sentry_apps/details.py
