+++ src/sentry/api/endpoints/integrations/sentry_apps/installation/details.py
