+++ src/sentry/api/endpoints/integrations/sentry_apps/installation/external_issue/index.py
