+++ src/sentry/api/endpoints/integrations/sentry_apps/internal_app_token/details.py
