+++ src/sentry/api/endpoints/integrations/sentry_apps/stats/details.py
